requirements.txt
# Core dependencies for SNP Diffusion package
torch>=1.12.0
torchvision>=0.13.0
torchmetrics>=1.7.4
numpy>=1.21.0
scipy>=1.7.0
diffusers==0.34.0
torch-fidelity>=0.3.0
scvi-tools>=1.3.3
reformer_pytorch>=1.4.4

# Data handling and processing
pandas>=1.3.0
h5py>=3.1.0
tqdm>=4.62.0

# Configuration and utilities
pyyaml>=5.4.0
ml-collections>=0.1.0
scikit-learn

# Visualization and plotting
matplotlib>=3.4.0
seaborn>=0.11.0
ipykernel

# Experiment tracking (optional)
neptune-client>=0.16.0

# Development and testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

wrapt>=1.17.3
magic-impute